// client/src/app/enum/global.enum.ts
export enum GLOBAL {
    URL_API = 'http://localhost:3977/api/',
    PATH_RESOURCE = '/assets/resource/',
    PATH_AUDIO = '/assets/audio/',
    PATH_JS = '/assets/js/',
    PATH_MAP = '/assets/resource/map/',
    PATH_TILEDSET = '/assets/resource/tiledset/',
    PATH_PLAYER = '/assets/resource/player/',
    CANVAS_ID = 'phaser-game'    
}